Update and rename messageReactionAddEvent.js to messageReactionAdd.js
